% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Init.R
\name{as.InferN0Scope}
\alias{as.InferN0Scope}
\title{Alias of infern0Init for conversion operation}
\usage{
as.InferN0Scope(object)
}
\arguments{
\item{versatile}{input data, matrix with gene/feature as rows and cells as collunms, seurat object or SingleCellExperiment object}
}
\value{
InferN0 scope
}
\description{
Alias of infern0Init for conversion operation
}
