% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{.infern0FitNegativeBinomial}
\alias{.infern0FitNegativeBinomial}
\title{Fit Negative Binomial to data}
\usage{
.infern0FitNegativeBinomial(scope)
}
\arguments{
\item{Infern0}{scope}
}
\description{
Fit Negative Binomial to data
}
