% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{.infern0ReverseTCMinMTX}
\alias{.infern0ReverseTCMinMTX}
\title{Convert file list of dense matrices to 3 files for gene name, cell names and sparse data MTX format}
\usage{
.infern0ReverseTCMinMTX(inprefix, outprefix, do.log = F)
}
\arguments{
\item{path}{or list of paths to tab delimited text table}

\item{pathprefix:}{output prefix for the 3 output files}
}
\description{
Convert file list of dense matrices to 3 files for gene name, cell names and sparse data MTX format
}
