% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{.infern0Test}
\alias{.infern0Test}
\title{Unsupported function for internal testing purposes}
\usage{
.infern0Test(scope)
}
\arguments{
\item{data}{matrix with gene/feature as rows and cells as collunms}
}
\description{
Unsupported function for internal testing purposes
}
