% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_cmpModeledVariance}
\alias{.infernal_cmpModeledVariance}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_cmpModeledVariance(infscope, list)
}
\description{
Internal Function, see R code for documentation instead
}
