% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_cmpVariance}
\alias{.infernal_cmpVariance}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_cmpVariance(infscope, list)
}
\description{
Internal Function, see R code for documentation instead
}
