% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_computeCovar}
\alias{.infernal_computeCovar}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_computeCovar(scptrexp, list)
}
\description{
Internal Function, see R code for documentation instead
}
