% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_convertDenseTextMatrixToSparseFormat}
\alias{.infernal_convertDenseTextMatrixToSparseFormat}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_convertDenseTextMatrixToSparseFormat(list)
}
\description{
Internal Function, see R code for documentation instead
}
