% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_createScope}
\alias{.infernal_createScope}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_createScope()
}
\description{
Internal Function, see R code for documentation instead
}
