% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_deserialize}
\alias{.infernal_deserialize}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_deserialize(obj, src)
}
\description{
Internal Function, see R code for documentation instead
}
