% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_exportpcs}
\alias{.infernal_exportpcs}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_exportpcs(infscope, rlist)
}
\description{
Internal Function, see R code for documentation instead
}
