% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_find_DE}
\alias{.infernal_find_DE}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_find_DE(scope, list)
}
\description{
Internal Function, see R code for documentation instead
}
