% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_find_markers}
\alias{.infernal_find_markers}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_find_markers(input)

.infernal_find_markers(input)
}
\description{
Internal Function, see R code for documentation instead

Internal Function, see R code for documentation instead
}
