% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_fitNegativeBinomial}
\alias{.infernal_fitNegativeBinomial}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_fitNegativeBinomial(data)
}
\description{
Internal Function, see R code for documentation instead
}
