% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_getDataRow}
\alias{.infernal_getDataRow}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_getDataRow(rscp, list)
}
\description{
Internal Function, see R code for documentation instead
}
