% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_getParameter}
\alias{.infernal_getParameter}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_getParameter(rscp, which)
}
\description{
Internal Function, see R code for documentation instead
}
