% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_getParameterList}
\alias{.infernal_getParameterList}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_getParameterList(rscp)
}
\description{
Internal Function, see R code for documentation instead
}
