% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_IdentifyConstrainedCovar}
\alias{.infernal_IdentifyConstrainedCovar}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_IdentifyConstrainedCovar(input, list, traincov, trainmean)
}
\description{
Internal Function, see R code for documentation instead
}
