% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_IdentifyNetwork}
\alias{.infernal_IdentifyNetwork}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_IdentifyNetwork(input, list, traincov, trainmean)
}
\description{
Internal Function, see R code for documentation instead
}
