% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_mergeSparseMatrices}
\alias{.infernal_mergeSparseMatrices}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_mergeSparseMatrices(matrixA, matrixB)
}
\description{
Internal Function, see R code for documentation instead
}
