% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_modelhidden_v2}
\alias{.infernal_modelhidden_v2}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_modelhidden_v2(infscope, data)
}
\description{
Internal Function, see R code for documentation instead
}
