% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_readDenseMatrix}
\alias{.infernal_readDenseMatrix}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_readDenseMatrix(scope, x, isRaw)
}
\description{
Internal Function, see R code for documentation instead
}
