% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_readMatrixFolder}
\alias{.infernal_readMatrixFolder}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_readMatrixFolder(list)
}
\description{
Internal Function, see R code for documentation instead
}
