% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_readScope}
\alias{.infernal_readScope}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_readScope(path)
}
\description{
Internal Function, see R code for documentation instead
}
