% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_readSparseMatrix}
\alias{.infernal_readSparseMatrix}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_readSparseMatrix(scope, x, isRaw)
}
\description{
Internal Function, see R code for documentation instead
}
