% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_reverseTCM}
\alias{.infernal_reverseTCM}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_reverseTCM(rscp, list)
}
\description{
Internal Function, see R code for documentation instead
}
