% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_reverseTCM_inMTX}
\alias{.infernal_reverseTCM_inMTX}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_reverseTCM_inMTX(list)
}
\description{
Internal Function, see R code for documentation instead
}
