% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_runTestFunction}
\alias{.infernal_runTestFunction}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_runTestFunction(data)
}
\description{
Internal Function, see R code for documentation instead
}
