% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_saveHierarchical}
\alias{.infernal_saveHierarchical}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_saveHierarchical(data, path)
}
\description{
Internal Function, see R code for documentation instead
}
