% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_setParameter}
\alias{.infernal_setParameter}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_setParameter(rscp, which, value)
}
\description{
Internal Function, see R code for documentation instead
}
