% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_wilcox_geneset}
\alias{.infernal_wilcox_geneset}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_wilcox_geneset(scope, list)
}
\description{
Internal Function, see R code for documentation instead
}
