% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_wilcox_matrix}
\alias{.infernal_wilcox_matrix}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_wilcox_matrix(input, list)
}
\description{
Internal Function, see R code for documentation instead
}
