% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_writeScope}
\alias{.infernal_writeScope}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_writeScope(infscope, path)
}
\description{
Internal Function, see R code for documentation instead
}
