% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernal_writeScope2}
\alias{.infernal_writeScope2}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernal_writeScope2(infscope, serialized, path)
}
\description{
Internal Function, see R code for documentation instead
}
