% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.infernalDensity}
\alias{.infernalDensity}
\title{Internal Function, see R code for documentation instead}
\usage{
.infernalDensity(data, inlist)
}
\description{
Internal Function, see R code for documentation instead
}
