% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.internal_Infern0_hierarchicalClustering_CustomGroups}
\alias{.internal_Infern0_hierarchicalClustering_CustomGroups}
\title{Internal Function, see R code for documentation instead}
\usage{
.internal_Infern0_hierarchicalClustering_CustomGroups(data, cellgroup,
  nbclusters)
}
\description{
Internal Function, see R code for documentation instead
}
