\name{InferN0-package}
\alias{InferN0-package}
\alias{InferN0}
\docType{package}
\title{
\packageTitle{InferN0}
}
\description{
\packageDescription{InferN0}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{InferN0}
\packageIndices{InferN0}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{InferN0}

Maintainer: \packageMaintainer{InferN0}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
