% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0_SaveAsCDTfile}
\alias{infern0_SaveAsCDTfile}
\title{Export hierarchical clustering(s) to target directory}
\usage{
infern0_SaveAsCDTfile(ifrn0, path)
}
\arguments{
\item{path}{to directory for output tree files}

\item{Infern0}{scope}
}
\description{
Export hierarchical clustering(s) to target directory
}
