% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Core.R
\name{infern0ComputeCovariance}
\alias{infern0ComputeCovariance}
\title{Estimate Covariance from sparse observations}
\usage{
infern0ComputeCovariance(i0scp, gene.list, method = "EM",
  cell.list = c(), cell.meta = c(), cell.cluster = "",
  nb.threads = 4, nb.crosseval.partition = 0, outlier_frequency = 0,
  EM.itemax = 10, seed = -1)
}
\arguments{
\item{i0scp}{Infern0 scope (or Input matrix gene as Rows, cells as Collumns)}

\item{gene.list}{List of transcript name for which a network is to be found (<50 gene expected)}

\item{method}{One for the following:
\enumerate{
  \item "Zero" computes covariance with 0 for missing values
  \item "Partial" computes partial-correlation matrix (can be singular!)
  \item "CorrectedPartial" computes partial-correlation matrix and ensures matrices are positive definite
  \item "EM" use inputation to find covariance
  \item "EMM" use inputation for conditionnal parameters to find covariance
  \item "Model" uses infered hidden states to normalize values (which assigns a non-trivial values to zero-counts)
}}

\item{cell.list}{Filter to listed cells}

\item{cell.meta}{Filter based on meta data associated to cells}

\item{cell.cluster}{uses matching clusterID as 'cell.list', ignored if custom cell.list is also provided}

\item{nb.threads}{number of threads required for compututations (>=1)}

\item{nb.crosseval.partition}{additionnal output which are multivariate sample covariance for (k-1)/k of the observations}

\item{outlier_frequency}{fraction of the data ignored as outliers (method in {"EM","EMM"} only)}

\item{EM.itemax}{number of EM iteration (method in {"EM","EMM"} only)}

\item{seed}{seed for randomly assigning cells into training and test set partitions}
}
\value{
list(Covar = matrix, Mean = vector, Input = matrix, Training.Covar = list(matrix()), Training.Mean = list(vector()), Testset.Partition = list(partition_index))
}
\description{
Estimate Covariance from sparse observations, which needs to account for missing information that can induce singular matrices if partial correlations are considered. Missing information can be inputed, or eigen values can be altered. Finally, one can infer the probability of biological dropout by modeling observations, and then normalize observation by reporting deviation to expectation instead.
}
