% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_IO.R
\name{infern0ConvertFilesToSparse}
\alias{infern0ConvertFilesToSparse}
\title{Convert file list of dense matrices to 3 files for gene name, cell names and sparse data MTX format}
\usage{
infern0ConvertFilesToSparse(pathlist, pathprefix, cellnameprefix = c(),
  is_floatingpoint = F, do.unzip = F, is_commaseparated = F)
}
\arguments{
\item{path}{or list of paths to tab delimited text table}

\item{pathprefix:}{output prefix for the 3 output files}

\item{cellnameprefix:}{string appended to cellnames from their respective file}

\item{is_floatingpoint:}{if set, expects floating point values}

\item{do.unzip:}{if set, expects the files to have an addionnnal .gz suffix, and temporalily uncompress them}
}
\description{
Convert file list of dense matrices to 3 files for gene name, cell names and sparse data MTX format
}
