% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_IO.R
\name{infern0ConvertToSeurat}
\alias{infern0ConvertToSeurat}
\title{Convert infern0 scope to Seurat Object, including custom normalization}
\usage{
infern0ConvertToSeurat(ifrn0, do.compute.TSNE = F)
}
\arguments{
\item{path}{Infern0 scope file path}

\item{do.hidden.as.pca:}{uses cell hidden.state to populate sro@dr$pca}
}
\value{
seurat Object
}
\description{
Convert infern0 scope to Seurat Object, including custom normalization
}
