% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_IO.R
\name{infern0ConvertToSingleCellExperiment}
\alias{infern0ConvertToSingleCellExperiment}
\title{Convert infern0 scope to Singlecell Experiement Object, including custom normalization}
\usage{
infern0ConvertToSingleCellExperiment(ifrn0)
}
\arguments{
\item{path}{Infern0 scope file path}

\item{do.hidden.as.pca:}{uses cell hidden.state to populate sro@dr$pca}
}
\value{
seurat Object
}
\description{
Convert infern0 scope to Singlecell Experiement Object, including custom normalization
}
