% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0DElistCompare}
\alias{infern0DElistCompare}
\title{Compute Wilcox test adjusted for missing data (Hypergeometric-Normal mixture for rank sum test)}
\usage{
infern0DElistCompare(value, pvalue, pvaluethreshold = 0.05)
}
\arguments{
\item{value:}{Matrix with measured feature that is to be ranked}

\item{pvalue:}{pvalue associated with observation}
}
\description{
Compute Wilcox test adjusted for missing data (Hypergeometric-Normal mixture for rank sum test)
}
