% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0ExportPrincipalComponents}
\alias{infern0ExportPrincipalComponents}
\title{Function that generates synthetic read counts, assuming there are "nb.hidden.col" cell types with "nb.hidden.row" subset of differentially expressed genes(or features)}
\usage{
infern0ExportPrincipalComponents(ifrn0, nb.pc = 50, nb.thread = 4)
}
\arguments{
\item{ifrn0}{InferN0 scope}

\item{nb.row}{number of features}

\item{nb.col}{number of cells}

\item{nb.hidden.row}{number of differentially expressed features groups}

\item{nb.hidden.col}{number of cell types}
}
\description{
Function that generates synthetic read counts, assuming there are "nb.hidden.col" cell types with "nb.hidden.row" subset of differentially expressed genes(or features)
}
