% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Core.R
\name{infern0FindCoexpressed}
\alias{infern0FindCoexpressed}
\title{Find Coexpressed}
\usage{
infern0FindCoexpressed(i0scp, gene.list, cell.list = as.character(c()),
  cell.cluster = "", nb.threads = 4, nb.output = 0,
  individual.coexp.max = 0, mincell = 5, do.append = F,
  call.findnetwork = c())
}
\arguments{
\item{i0scp}{Infern0 scope}

\item{gene.list}{List of transcript name for which a network is to be found (<50 gene expected)}

\item{cell.list}{Filter to listed cells}

\item{cell.cluster}{uses matching clusterID as 'cell.list', ignored if custom cell.list is also provided}

\item{nb.threads}{number of threads required for compututations (>=1)}

\item{nb.output}{number of coexpressed transcript to find (default: same as size of 'gene.list')}

\item{mincell}{minimum amount of cell with transcript pairs needed to evalutate the correlation.}

\item{do.call.findnetwork}{list of arguments supplied to a call to 'infern0IdentifyNetwork' done uppon completion, which overwrites output.}
}
\description{
Compute Partial correlation assoiated with a subset of all transcripts
}
