% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0FindConstrainedCovariance}
\alias{infern0FindConstrainedCovariance}
\title{Find optimal Covariance with associated constrained Precision matrix}
\usage{
infern0FindConstrainedCovariance(constraint, covarstr)
}
\arguments{
\item{symmetric}{sparse matrix constraint}

\item{list}{equivalent to what is outputed by the 'infern0ComputeCovariance' function (if used, genes must agree)}
}
\value{
list
}
\description{
Find optimal Covariance with associated constrained Precision matrix
}
