% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0FindMarkers}
\alias{infern0FindMarkers}
\title{Compute Wilcox test adjusted for missing data (Hypergeometric-Normal mixture for rank sum test)}
\usage{
infern0FindMarkers(input, nbpart = 1, partition = c(),
  ordering = c(), print.progress = T, do.quartile.average = T,
  is.ordering.a.partition = F, use.pure.wilcox = F,
  do.downsample = F)
}
\arguments{
\item{input.matrix:}{InferN0 Scope with transposed matrix...}

\item{partition:}{list of positive integer with cluster identity of cells}

\item{ordering:}{cells ordered by major counfounding factor, used to partition into quartiles}

\item{is.ordering.a.partition:}{defines the partition directly}

\item{use.pure.wilcox:}{use pure wilcox test}
}
\description{
Compute Wilcox test adjusted for missing data (Hypergeometric-Normal mixture for rank sum test)
}
