% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Plot.R
\name{infern0FrequencyBarPlot}
\alias{infern0FrequencyBarPlot}
\title{Plot for proportion/frequency for pair for meta data entries}
\usage{
infern0FrequencyBarPlot(xannot, hannot, wannot = c(),
  do.horizscale = T, use.colors = c(),
  show.label.lowerbound.fraction = 0.05, do.resize = c(),
  is.vertical = F, plot.attribs = c(), return.counts.instead = F,
  make.piechart = F)
}
\arguments{
\item{ggplot.or.plot.attrib:}{Either a ggplot object or a default list of attributes that is to be modified (effectively equivalent to setting the "baseAttribs" instead). If a ggplot is provided, outputs an ggplot object with the applied modifications instead.}

\item{baseAttribs:}{list of attributes that is to be modified}
}
\description{
Plot for proportion/frequency for pair for meta data entries
}
