% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0GenerateSyntheticData}
\alias{infern0GenerateSyntheticData}
\title{Function that generates synthetic read counts, assuming there are "nb.hidden.col" cell types with "nb.hidden.row" subset of differentially expressed genes(or features)}
\usage{
infern0GenerateSyntheticData(nb.row = 320, nb.col = 200,
  nb.hidden.row = 8, nb.hidden.col = 5, mixedstate_weight = 1)
}
\arguments{
\item{nb.row}{number of features}

\item{nb.col}{number of cells}

\item{nb.hidden.row}{number of differentially expressed features groups}

\item{nb.hidden.col}{number of cell types}
}
\value{
InferN0 scope
}
\description{
Function that generates synthetic read counts, assuming there are "nb.hidden.col" cell types with "nb.hidden.row" subset of differentially expressed genes(or features)
}
