% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0GenerateSyntheticNetworkData}
\alias{infern0GenerateSyntheticNetworkData}
\title{Generate Synthetic Data with a given Network}
\usage{
infern0GenerateSyntheticNetworkData(Z, n = 100, mu.min = 10,
  mu.max = 10, p.min = 0.3, p.max = 0.3, s.min = 3, s.max = 3,
  noise.min = 0, noise.max = 0, muvec = c(), pvec = c(),
  svec = c(), noisevec = c(), do.log.transform = T,
  is.log.dependency = T)
}
\arguments{
\item{mu.min, mu.max:}{mean expression of transcripts}

\item{p.min, p.max:}{dropout rate transcripts}

\item{s.min, s.max:}{background noise for expression of transcripts}

\item{Z:}{Symetrix matrix as network}

\item{n:}{number of generated cells (collumns)}
}
\description{
Generate Synthetic Data with a given Network
}
