% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Query.R
\name{infern0Get}
\alias{infern0Get}
\title{Query variable within InferN0Scope}
\usage{
infern0Get(scope, varname)
}
\arguments{
\item{Infern0}{scope}

\item{Infern0}{variable name, one of ("cell.cluster", "gene.cluster", "cell.scale", "cell.order","gene.scale","gene.order")}
}
\value{
requested attribute of Infern0 scope
}
\description{
Query variable within InferN0Scope
}
