% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0GetDensity}
\alias{infern0GetDensity}
\title{Example function to analyse data using provided function. This function serves as a guideline on the intended use of other functions.}
\usage{
infern0GetDensity(coor, weight = c(), data = c(), covariance = c(),
  bandwidth = 0.25, mapsize = 256, rect = c(),
  do.normalize.bandwidth = T)
}
\arguments{
\item{2-collumn}{data points}

\item{network_genelist}{list of genes for network inference}

\item{nbgene.to.detect}{number of co-expressed genes to detect, ("default" is a number equal to the length of 'network_genelist')}
}
\description{
Example function to analyse data using provided function. This function serves as a guideline on the intended use of other functions.
}
