% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0HierarchicalClustering}
\alias{infern0HierarchicalClustering}
\title{Execute HierarchicalClustering}
\usage{
infern0HierarchicalClustering(ifrn0, cluster.cells = TRUE,
  write.ordering = TRUE)
}
\arguments{
\item{cluster.cells}{, TRUE to  cluster cell, FALSE to cluster GENES}

\item{Infern0}{scope}

\item{(over)writes}{in Infern0 scope the ordering as the new hierarchical clustering indicates}
}
\description{
Execute HierarchicalClustering
}
