% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0HierarchicalClusteringWithInput}
\alias{infern0HierarchicalClusteringWithInput}
\title{Execute HierarchicalClustering}
\usage{
infern0HierarchicalClusteringWithInput(ifrn0, cells.group)
}
\arguments{
\item{Infern0}{scope}
}
\description{
Execute HierarchicalClustering
}
