% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Core.R
\name{infern0IdentifyNetworkOlderVersion}
\alias{infern0IdentifyNetworkOlderVersion}
\title{Execute Network inference Older Version...}
\usage{
infern0IdentifyNetworkOlderVersion(i0scp, gene.list = c(""),
  method = "EM", evaltarget = c(), nb.crosseval.partition = 8,
  cell.list = as.character(c()), cell.cluster = "", nb.threads = 4,
  max.cyclic.component = 10, label.cex = 0.75, edge.width = 10,
  do.plot = T, vertex.size.base = 8, vertex.size.incr = 1.6,
  vertex.height = 5, check.nb = 10, check.min = 0,
  color = "#FF4400", do.manually.select.nb.edges = T)
}
\arguments{
\item{i0scp}{Infern0 scope (or list equivalent to what is outputed by the 'infern0ComputeCovariance' function (if used, ignores any provided arguments that shared with that function)}

\item{list}{of transcript name for which a network is to be found (<50 gene expected)}

\item{method:}{string argument needed to compute covariance ("EM", "EMM","Zero", "Partial", "CorrectedPartial","Model" see the 'infern0ComputeCovariance' function)}

\item{evaltarget:}{optionnal 'true' Network argument to report on true/false prediction rates}

\item{max.cyclic.component:}{maximal size for a cyclic compnent (cycles in network), search might be slower if a more larger threshold size}

\item{cell.list:}{list of cell names, list of cell index or T/F filter vector}

\item{cell.cluster:}{uses matching clusterID as 'cell.list', ignored if custom cell.list is also provided}

\item{nb.threads:}{number of threads required for compututations (>=1)}

\item{check.nb:}{upperbound on number of edges for the search depth, beyond the optimum reported.}

\item{check.min:}{lowerbound on number of edges for the search depth. (if =0, sets value as size of network instead)}

\item{do.manually.select.nb.edges:}{uses 'check.nb' as forced number of edges in output graph}
}
\description{
Execute Network inference Older Version...
}
