% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Core.R
\name{infern0IdentifyNetworkPipeline}
\alias{infern0IdentifyNetworkPipeline}
\title{Effectively execute 'infern0FindCoexpressed', 'infern0IdentifyNetwork' and 'infern0PlotNetwork' while returning the output of 'infern0IdentifyNetwork' with default parameters}
\usage{
infern0IdentifyNetworkPipeline(i0scp, gene.list,
  cell.list = as.character(c()), cell.cluster = "", method = "EMM",
  nb.threads = 4, max.cyclic.component = 10, label.cex = 0.65,
  edge.width = 10, do.plot = T, vertex.size.base = 8,
  vertex.size.incr = 1.6, vertex.height = 5, check.nb = 10,
  check.min = 0, color = "#FF4400", extendedcolor = "#FFFFFFFF",
  nb.output = 0, individual.coexp.max = 0, mincell = 5)
}
\arguments{
\item{i0scp}{Infern0 scope}
}
\description{
Sequentially execute 'infern0FindCoexpressed', 'infern0IdentifyNetwork', 'infern0PlotNetwork'.
}
