% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Init.R
\name{infern0Init}
\alias{infern0Init}
\title{Initialization of InferN0Scope}
\usage{
infern0Init(input, meta.data = c(), use.normalized.data = F)
}
\arguments{
\item{input}{versatile input data, matrix with gene/feature as rows and cells as collunms, seurat object or SingleCellExperiment object}

\item{meta.data}{meta data associated with each cell}

\item{use.normalized.data}{selects if the input populates the rawdata (integer counts) or the data slot for normalized data (floating-point value)}
}
\value{
InferN0 scope object (can only be saved to a file using \code{infern0SaveFile})
}
\description{
Initialize Infern0 scope from input data, which can be a matrix, a singlecelleperiment or Seurat object
}
