% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0internal_FilterRowCol}
\alias{infern0internal_FilterRowCol}
\title{Internal: filter maximimal numb of row/col to retain percentage of data}
\usage{
infern0internal_FilterRowCol(input, fraction = 0.9, cost_ratio = -1,
  do_print = FALSE)
}
\arguments{
\item{SingleCellExperiement}{Object to load}
}
\description{
Internal: filter maximimal numb of row/col to retain percentage of data
}
