% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_IO.R
\name{infern0LoadFileOld}
\alias{infern0LoadFileOld}
\title{Load Infern0 Scope}
\usage{
infern0LoadFileOld(path = "./ifrn0.scp")
}
\arguments{
\item{path}{Infern0 scope file path}
}
\value{
InferN0 scope (can only be saved to a file using \code{infern0SaveFile})
}
\description{
Loads Infern0 scope from file.
}
