% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Core.R
\name{infern0ModelData}
\alias{infern0ModelData}
\title{Sparse Factor based Negative-Binomial modeling}
\usage{
infern0ModelData(i0scp, nb.hidden.row = 8, nb.hidden.col = 5,
  nb.step = 200, nb.thread = 4, cell.cluster = c(),
  do.output.tiff = FALSE, version = 1, state.labels = c("CELL_STATE",
  "GENE_STATE"), do.runTSNE = T, use.heavy.prior = F)
}
\arguments{
\item{i0scp}{InferN0 scope}

\item{nb.hidden.row}{number of differentially expressed features groups}

\item{nb.hidden.col}{number of cell types}

\item{nb.step}{number of EM steps to perform}

\item{nb.thread}{number of threads}

\item{cell.cluster}{(optionnal) start search using input clusters}

\item{state.labels}{pair of names or name prefix used for outputs (stored within i0scp)}

\item{do.runTSNE}{T/F flag for runing Rtsne of hidden states}
}
\description{
Function that normalizes the data by recovering the major coexpression factors accross cells (aka cell-types) and genes
}
