% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Plot.R
\name{infern0Overlay}
\alias{infern0Overlay}
\title{Function used to easily populate ggplot related options stored in a list, that can be used by ploting function "overlay" as plot.attrib argument}
\usage{
infern0Overlay(infrnscp, query, transform = c(), coor = "TSNE",
  filter = c(), filter.meta = c(), plot.attribs = c(),
  color.palette = "fireice", na.alpha = 0.1, na.color = "#888888")
}
\arguments{
\item{infrnscp:}{InferN0 scope}

\item{query:}{meta.data collumn or gene}

\item{transform:}{string representing a transformation of input field}

\item{coor:}{1 to 3 strings representing x,y,z coordinates for overlay. If a single string is provided, a pair with "_0" and "_1" is used instead.}

\item{filter:}{T/F vector to select cells or genes to render}

\item{filter.meta:}{meta.data/gene.meta filter criterions to select cells/genes to render}

\item{plot.attribs:}{additionnal flags to change ggplot layout, use the output from infern0plotAttribs}

\item{color.palette:}{list of colors for overlays if the query is a numerical value}
}
\value{
ggplot2 object or rgl overlay
}
\description{
Function used to easily populate ggplot related options stored in a list, that can be used by ploting function "overlay" as plot.attrib argument
}
\examples{
scp <- infern0Init(datamatrix)
infern0Overlay(scp, "geneA", coor= c("geneB", "geneC"))
}
