% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Plot.R
\name{infern0PlotNetwork}
\alias{infern0PlotNetwork}
\title{Plot Inferred Network}
\usage{
infern0PlotNetwork(out, layout = c(), nb.edges = 0,
  vertex.size.base = 12, vertex.size.incr = 2.4, edge.width = 10,
  edge.color = "#000000FF", edge.color.low = "#222222FF",
  vertex.height = 5, label.cex = 0.75, incl_isolates = F,
  color = "#FF4400", do.make.interactive = T)
}
\arguments{
\item{networkoutput:}{list generated by 'Infern0IdentifyNetwork'}
}
\description{
Plot Inferred Network
}
