% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_IO.R
\name{infern0ReadRangerMatrixCPP}
\alias{infern0ReadRangerMatrixCPP}
\title{Load read counts using files produced by 'cellranger count (...)'}
\usage{
infern0ReadRangerMatrixCPP(path, threshold.low.UMI = 0,
  threshold.low.GENE = 0,
  threshold.genes.ignored = c("thisgenedoesnotexistsurely"),
  do.soup.regression = F, do.print.progress = T, do.plot.soup = F,
  plot.soup.title = c(), threshold.soup.fraction = 1,
  threshold.soup.logUMIslope = 0, is.floatingpoint = F,
  do.TCM.reverse = F, do.output.data.only = F, nb.soup.samples = 32)
}
\arguments{
\item{directory}{path to a cellranger output folder, which expected that contains 3 files named barcodes.tsv genes.tsv matrix.mtx respectively. (ex: "./data/")}

\item{colsubset:}{if set, getCells with matching name only}
}
\value{
List containing the following: (data, filtered.data, etc)
}
\description{
Load read counts using files produced by 'cellranger count (...)'
}
