% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0ReverseSizeFactor}
\alias{infern0ReverseSizeFactor}
\title{Recover Raw Counts}
\usage{
infern0ReverseSizeFactor(infrn0, do.log = F)
}
\arguments{
\item{path}{or list of paths to tab delimited text table}

\item{pathprefix:}{output prefix for the 3 output files}
}
\description{
Recovers the 'raw.data' using 'data' assuming cells had their expression normalized (no transcript-length/gene specific normalization)
}
