% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0RunExample}
\alias{infern0RunExample}
\title{Example function to analyse data using provided function. This function serves as a guideline on the intended use of other functions.}
\usage{
infern0RunExample(data, network_genelist,
  nbgene.to.detect = "use.default", nb.threads = 4)
}
\arguments{
\item{network_genelist}{list of genes for network inference}

\item{nbgene.to.detect}{number of co-expressed genes to detect, ("default" is a number equal to the length of 'network_genelist')}

\item{Seurat}{Object/SingleCellExperiment/data matrix with gene/feature as rows and cells as collunms}
}
\description{
Example function to analyse data using provided function. This function serves as a guideline on the intended use of other functions.
}
