% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_IO.R
\name{infern0SaveFile}
\alias{infern0SaveFile}
\title{Save Infern0 Scope}
\usage{
infern0SaveFile(ifrn0, path = "./ifrn0.scp")
}
\arguments{
\item{path}{Infern0 scope file path}
}
\value{
void
}
\description{
Saves Infern0 scope to file. (Needed Previously generated by "loadInferN0file")
}
