% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0SaveHierarchical}
\alias{infern0SaveHierarchical}
\title{(...)}
\usage{
infern0SaveHierarchical(scope, pathprefix)
}
\arguments{
\item{Infern0}{scope}

\item{path}{prefix for output files (ex: "~/outdir/output")}
}
\description{
(...)
}
