% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Query.R
\name{infern0Set}
\alias{infern0Set}
\title{Set variable within InferN0Scope}
\usage{
infern0Set(scope, varname, value)
}
\arguments{
\item{value}{to set, its type must match the output of 'Infern0Get' with matching variable name}

\item{Infern0}{scope}

\item{Infern0}{variable name, one of ("cell.cluster", "gene.cluster", "cell.scale", "cell.order","gene.scale","gene.order")}
}
\value{
requested attribute of Infern0 scope
}
\description{
Set variable within InferN0Scope
}
