% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFRN_Query.R
\name{infern0SetClusterIdentity}
\alias{infern0SetClusterIdentity}
\title{(...)}
\usage{
infern0SetClusterIdentity(scope, assosication)
}
\arguments{
\item{Infern0}{scope}

\item{list/matrix}{with clusterID as values}
}
\description{
(...)
}
