% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0SparseMatrixMerge}
\alias{infern0SparseMatrixMerge}
\title{Merge sparse matrices using matching row/column names}
\usage{
infern0SparseMatrixMerge(matrixA, matrixB, mode = "o")
}
\arguments{
\item{matrixA:}{first matrix to merge}

\item{matrixB:}{second matrix to merge}

\item{mode:}{operation "o" overwrite non-zero, "O" overwrite}
}
\description{
Merge sparse matrices using matching row/column names
}
