% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0Summary}
\alias{infern0Summary}
\title{Displays a summary of a Infern0 scope}
\usage{
infern0Summary(ifrn0)
}
\arguments{
\item{path}{Infern0 scope file path}
}
\value{
void
}
\description{
Displays a summary of a Infern0 scope
}
