% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_functions.R
\name{infern0ZeroCorrectedWilcox}
\alias{infern0ZeroCorrectedWilcox}
\title{Compute Wilcox test adjusted for missing data (Hypergeometric-Normal mixture for rank sum test)}
\usage{
infern0ZeroCorrectedWilcox(input, list.pos, list.neg, do.transpose = T,
  do.cmp.mean = F, use.clusterID = F, purewilcox = F,
  print.progress = T)
}
\arguments{
\item{input.matrix:}{sparse matrix or InferN0 Scope with transposed matrix...}

\item{list.pos:}{list of rows for Negative Partition, or T/F vector matching dimention}

\item{list.neg:}{list of rows for Positive Partition, or T/F vector matching dimention}
}
\description{
Compute Wilcox test adjusted for missing data (Hypergeometric-Normal mixture for rank sum test)
}
